feat(run): refactor monolithic run.go into focused components
- Split run.go into: paths, remote (fetch + decision), render, link, exec, target, variables, checksum, commit_hash, sync, clean
- Introduce remote fetch decision (decideRemoteFetch) and modular ensureRendered/linkRendered/pruneOldRendered
- Add dedicated test files: render_test, link_test, exec_test, remote_fetch_test, remote_decision_test, sync_pipeline_test, checksum_test, commit_hash_test, variables_test, clean_test, paths_test, unit_test (misc)
- Remove legacy run_* test files with migrated coverage
- Maintain existing behavior (no functional CLI changes); improved test isolation and maintainability

Refs: architecture discussion in PR description
