+++ internal/run/variables_test.go
@@ -71,45 +71,43 @@ func TestResolveVariablesMissingFile(t *testing.T) {
 // TestAllowMissingVsStrict contrasts strict (missing variable => error) vs
 // allowMissing (missing variable => empty string) template rendering modes.
 func TestAllowMissingVsStrict(t *testing.T) {
-	tmp := t.TempDir()
-	oldWd, _ := os.Getwd()
-	os.Chdir(tmp)
-	defer os.Chdir(oldWd)
-	// template with missing VAR2
-	templateSrc := filepath.Join(tmp, "src")
-	os.MkdirAll(templateSrc, 0o755)
-	os.WriteFile(filepath.Join(templateSrc, "f.tpl"), []byte("A={{ .VAR1 }} B={{ .VAR2 }}"), 0o644)
-	// stub clone
-	origClone := cloneFunc
-	cloneFunc = func(repo, ref, cacheDir string) (string, error) {
-		dst := filepath.Join(cacheDir, "repo")
-		os.MkdirAll(dst, 0o755)
-		b, _ := os.ReadFile(filepath.Join(templateSrc, "f.tpl"))
-		os.WriteFile(filepath.Join(dst, "f.tpl"), b, 0o644)
-		return dst, nil
-	}
-	defer func() { cloneFunc = origClone }()
-	// strict -> error
-	strictCfg := &config.DuckConf{Version: 1, Default: "build", Targets: map[string]config.Target{"build": {Binary: "echo", FileFlag: "-f", Template: config.Template{Repo: "stub", Path: "f.tpl"}, Variables: map[string]config.VarValue{"VAR1": config.NewLiteralVar("one")}}}}
-	if err := Sync(strictCfg, "default", false, defaultSecurityConfigVars(), nil, nil); err == nil {
-		t.Fatalf("expected error for missing variable in strict mode")
-	}
-	// allowMissing -> empty value
-	allowCfg := &config.DuckConf{Version: 1, Default: "build", Targets: map[string]config.Target{"build": {Binary: "echo", FileFlag: "-f", Template: config.Template{Repo: "stub", Path: "f.tpl", AllowMissing: true}, Variables: map[string]config.VarValue{"VAR1": config.NewLiteralVar("one")}}}}
-	// stub exec so Exec path safe
-	origExec := execCommand
-	execCommand = func(name string, args ...string) *exec.Cmd { return exec.Command("true") }
-	defer func() { execCommand = origExec }()
-	if err := Sync(allowCfg, "build", false, defaultSecurityConfigVars(), nil, nil); err != nil {
-		t.Fatalf("allowMissing sync err: %v", err)
-	}
-	link := filepath.Join(".duck", "build", "f")
-	target, _ := os.Readlink(link)
-	if !filepath.IsAbs(target) {
-		target = filepath.Join(filepath.Dir(link), target)
-	}
-	b, _ := os.ReadFile(target)
-	if !strings.Contains(string(b), "B=") {
-		t.Fatalf("expected empty B= got %q", string(b))
-	}
+	withTempWD(t, func() {
+		// template with missing VAR2
+		templateSrc := filepath.Join("src")
+		os.MkdirAll(templateSrc, 0o755)
+		os.WriteFile(filepath.Join(templateSrc, "f.tpl"), []byte("A={{ .VAR1 }} B={{ .VAR2 }}"), 0o644)
+		// stub clone
+		origClone := cloneFunc
+		cloneFunc = func(repo, ref, cacheDir string) (string, error) {
+			dst := filepath.Join(cacheDir, "repo")
+			os.MkdirAll(dst, 0o755)
+			b, _ := os.ReadFile(filepath.Join(templateSrc, "f.tpl"))
+			os.WriteFile(filepath.Join(dst, "f.tpl"), b, 0o644)
+			return dst, nil
+		}
+		defer func() { cloneFunc = origClone }()
+		// strict -> error
+		strictCfg := &config.DuckConf{Version: 1, Default: "build", Targets: map[string]config.Target{"build": {Binary: "echo", FileFlag: "-f", Template: config.Template{Repo: "stub", Path: "f.tpl"}, Variables: map[string]config.VarValue{"VAR1": config.NewLiteralVar("one")}}}}
+		if err := Sync(strictCfg, "default", false, defaultSecurityConfigVars(), nil, nil); err == nil {
+			t.Fatalf("expected error for missing variable in strict mode")
+		}
+		// allowMissing -> empty value
+		allowCfg := &config.DuckConf{Version: 1, Default: "build", Targets: map[string]config.Target{"build": {Binary: "echo", FileFlag: "-f", Template: config.Template{Repo: "stub", Path: "f.tpl", AllowMissing: true}, Variables: map[string]config.VarValue{"VAR1": config.NewLiteralVar("one")}}}}
+		// stub exec so Exec path safe
+		origExec := execCommand
+		execCommand = func(name string, args ...string) *exec.Cmd { return exec.Command("true") }
+		defer func() { execCommand = origExec }()
+		if err := Sync(allowCfg, "build", false, defaultSecurityConfigVars(), nil, nil); err != nil {
+			t.Fatalf("allowMissing sync err: %v", err)
+		}
+		link := filepath.Join(".duck", "build", "f")
+		target, _ := os.Readlink(link)
+		if !filepath.IsAbs(target) {
+			target = filepath.Join(filepath.Dir(link), target)
+		}
+		b, _ := os.ReadFile(target)
+		if !strings.Contains(string(b), "B=") {
+			t.Fatalf("expected empty B= got %q", string(b))
+		}
+	})
 }
